[ALPS09800096] arp_mon: Update NUD state parameter type
[Description]
Changed the type of the NUD state parameter from uint64_t to uint8_t
in arpMonHandleNudState function to match the expected data type and
prevent potential issues.

[Test]
UT

MTK-Commit-Id: dc06caaf77acfe17b0fb12772b1c416811996975

CR-Id: ALPS09800096
Feature: Others
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/10548915
GAI-Code-Review: srv_ai_review001 <srv_ai_review001@mediatek.com>
AutoUT-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Commit-Check: srv_check_service <srv_check_service@mediatek.com>
Test: srv_pf_nep_sanity
ODB-Check: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Coverity-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Build: srv_preflight_b001 <srv_preflight_b001@mediatek.com>

+++ include/mgmt/arp_mon.h
@@ -84,7 +84,7 @@ u_int8_t arpMonIpIsCritical(struct ADAPTER *ad, struct MSDU_INFO *prMsduInfo);
 u_int8_t arpMonIsCritical(struct ADAPTER *ad, uint8_t ucBssIdx);
 void arpMonResetArpDetect(struct ADAPTER *ad, uint8_t ucBssIdx);
 void arpMonHandleLegacyBTOEvent(struct ADAPTER *ad);
-void arpMonHandleNudState(struct ADAPTER *prAdapter, uint64_t state,
+void arpMonHandleNudState(struct ADAPTER *prAdapter, uint8_t state,
 	uint8_t ucBssIndex);
 
 #if CFG_QM_ARP_MONITOR_MSG
